requirements.txt
requests~=2.32.5
numpy~=2.3.2
pillow~=11.3.0
mysql-connector-python